% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{gmd}
\alias{gmd}
\title{Constructor for Gaussian Mixture Data}
\usage{
gmd(sample, dictionary, compute_support = TRUE,
  compute_centring = TRUE)
}
\arguments{
\item{sample}{A data frame with 3 variables: \code{observation} which labels
each observation, \code{component} which labels each mixture component and
\code{mixing} which gives the weight of each component within each
observation.}

\item{dictionary}{A data frame with 3 variables: \code{component} which
labels each mixture component and \code{mean} and \code{precision} which
respectively give the mean and the precision (inverse variance) of the
Gaussian distribution of the corresponding component.}
}
\value{
An object of class \code{\link{gmd}}, which is effectively a list
  with 2 components: \code{sample} and \code{dictionary}, as described by the
  input arguments.
}
\description{
Constructor for Gaussian Mixture Data
}
\examples{
N <- 100
M <- 4
w <- matrix(runif(N * M), N, M)
w <- w / rowSums(w)
s <- tidyr::crossing(
  observation = paste0("O", 1:N),
  component = paste0("C", 1:M)
) \%>\%
dplyr::mutate(mixing = as.numeric(t(w)))
d <- tibble::tibble(
  component = paste0("C", 1:M),
  mean = numeric(M),
  precision = 1:M
)
x <- gmd(s, d)
}
