% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{mean.gmd}
\alias{mean.gmd}
\title{Sample Mean for Gaussian Mixture Data}
\usage{
\method{mean}{gmd}(x, trim = 0, na.rm = TRUE, K = NULL, rule = 2)
}
\arguments{
\item{x}{An object of class \code{\link{gmd}}.}

\item{native}{Boolean specyfing whether the resulting mean should be mapped
back as a density (default: \code{TRUE}) or if its crl-transform should be
return instead.}
}
\value{
The sample mean of the input Gaussian mixture data in density space
  (if \code{native==TRUE}) or in clr-space otherwise.
}
\description{
Sample Mean for Gaussian Mixture Data
}
\examples{
N <- 100
M <- 4
w <- matrix(runif(N * M), N, M)
w <- w / rowSums(w)
s <- tidyr::crossing(
  observation = paste0("O", 1:N),
  component = paste0("C", 1:M)
) \%>\%
dplyr::mutate(mixing = as.numeric(t(w)))
d <- tibble::tibble(
  component = paste0("C", 1:M),
  mean = numeric(M),
  precision = 1:M
)
x <- gmd(s, d)
mean(x)
}
