% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{mean.sgd}
\alias{mean.sgd}
\title{Mean for Single Gaussian Data}
\usage{
\method{mean}{sgd}(x, ref_mean = 0, ref_precision = 1, ...)
}
\arguments{
\item{x}{An object of class \code{\link{sgd}}.}

\item{ref_mean}{A scalar giving the mean of the Gaussian distribution used as
reference for the Bayes space (default: \code{0}).}

\item{ref_precision}{A scalar giving the precision (inverse of the variance)
of the Gaussian distribution used as reference for the Bayes space
(default: \code{1}). This should be a strictly positive value.}
}
\value{
A named length-2 numeric vector providing the \code{mean} and
  \code{precision} of the mean Gaussian distribution of the data set.
}
\description{
Mean for Single Gaussian Data
}
\examples{
x <- sgd(
  c(mean =  0, precision = 1  ),
  c(mean =  3, precision = 0.5),
  c(mean = -1, precision = 2)
)
mean(x)
}
