% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{plot.sgd}
\alias{plot.sgd}
\title{Plot for Single Gaussian Data}
\usage{
\method{plot}{sgd}(x, scale = 3L, centered = FALSE, add_mean = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{sgd}}.}

\item{scale}{An integer that defines the range of \code{x} values for
plotting as a multiple of the maximum observed standard deviation (default:
\code{3L}).}

\item{centered}{A boolean specifying whether individual densities should be
centered prior to plotting (default: \code{FALSE}).}
}
\value{
Invisibly returns the input object.
}
\description{
Plot for Single Gaussian Data
}
\examples{
x <- sgd(
  c(mean =  0, precision = 1  ),
  c(mean =  3, precision = 0.5),
  c(mean = -1, precision = 2  )
)
plot(x, centered = FALSE)
plot(x, centered = TRUE, xlab = "values")
}
