% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{sgd}
\alias{sgd}
\title{Constructor for Single Gaussian Data}
\usage{
sgd(...)
}
\arguments{
\item{...}{A number of named length-2 numeric vectors providing the
\code{mean} and \code{precision} of each Gaussian distribution in the data
set.}
}
\value{
An object of class \code{\link{sgd}} which is effectively a list of
  named length-2 numeric vectors providing the \code{mean} and
  \code{precision} of each Gaussian distribution in the data set.
}
\description{
Constructor for Single Gaussian Data
}
\examples{
x <- sgd(
  c(mean =  0, precision = 1  ),
  c(mean =  3, precision = 0.5),
  c(mean = -1, precision = 2  )
)
}
