% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.gmd}
\alias{[.gmd}
\title{Subset Operator for Gaussian Mixture Data}
\usage{
\method{[}{gmd}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{gmd}}.}

\item{i}{A valid expression giving instructions for subsetting observations.}

\item{j}{Unused.}

\item{...}{Unused.}

\item{drop}{Unused.}
}
\value{
An object of the same class as the input object with only the subset
  of observations identified by the expression in argument \code{i}.
}
\description{
Subset Operator for Gaussian Mixture Data
}
\examples{
N <- 100
M <- 4
w <- matrix(runif(N * M), N, M)
w <- w / rowSums(w)
s <- tidyr::crossing(
  observation = paste0("O", 1:N),
  component = paste0("C", 1:M)
) \%>\%
dplyr::mutate(mixing = as.numeric(t(w)))
d <- tibble::tibble(
  component = paste0("C", 1:M),
  mean = numeric(M),
  precision = 1:M
)
x <- gmd(s, d)
x[2]
x[2:4]
x[c(1, 3)]
x[-1]
x["O2"]
x[c("O1", "O3")]
}
